+++ R/Utils.R
@@ -322,11 +322,9 @@
         if ((grid_lons != common_grid_lons) || 
             (grid_lats != common_grid_lats) || 
             (grid_type != common_grid_type) ||
-            ((lon[1] != first_common_grid_lon) 
-             && !work_piece[['single_dataset']])) {
+            (lon[1] != first_common_grid_lon)) { 
           if (grid_lons == common_grid_lons && grid_lats == common_grid_lats &&
-              grid_type == common_grid_type && lon[1] != first_common_grid_lon &&
-              !work_piece[['single_dataset']]) {
+              grid_type == common_grid_type && lon[1] != first_common_grid_lon) {
             remove_shift <- TRUE
           }
           remap_needed <- TRUE
@@ -595,9 +593,6 @@
         nltime <- fnc$var[[namevar]][['dim']][[match(time_dimname, var_dimnames)]]$len
         expected_dims <- c(expected_dims, time_dimname)
         dim_matches <- match(expected_dims, var_dimnames)
-        first_time_step_in_file <- fnc$var[[namevar]][['dim']][[match(time_dimname,
-                                        var_dimnames)]]$vals[1]
-        time_units <- fnc$var[[namevar]][['dim']][[match(time_dimname, var_dimnames)]]$units
       } else {
         if (!is.null(old_members_dimname)) {
           expected_dims[which(expected_dims == 'lev')] <- old_members_dimname
@@ -638,7 +633,7 @@
         time_indices <- ts(time_indices, start = c(years[1], mons[1]), 
                            end = c(years[length(years)], mons[length(mons)]),
                            frequency = 12)
-                ltimes_list <- list()
+        ltimes_list <- list()
         for (sdate in work_piece[['startdates']]) {
           selected_time_indices <- window(time_indices, start = c(as.numeric(
                                    substr(sdate, 1, 4)), as.numeric(substr(sdate, 5, 6))), 
@@ -945,9 +940,7 @@
   if (explore_dims) {
     list(dims = dims, is_2d_var = is_2d_var, grid = grid_name, 
          units = units, var_long_name = var_long_name, 
-         data_across_gw = data_across_gw, array_across_gw = array_across_gw,
-         time_dim = list(first_time_step_in_file = first_time_step_in_file,
-                        time_units = time_units))
+         data_across_gw = data_across_gw, array_across_gw = array_across_gw)
   } else {
     ###if (!silent && !is.null(progress_connection) && !is.null(work_piece[['progress_amount']])) {
     ###  foobar <- writeBin(work_piece[['progress_amount']], progress_connection)