+++ R/zzz.R
@@ -1,0 +1,3 @@
+.onAttach <- function(libname, pkgname){
+  packageStartupMessage("Warning: This package has been deprecated. Please consider using the package 's2dv' instead.")
+}